Change default so "td200" means with status line OFF.
Now, to specify with status line on, people must use "td200-s".

This was changed because there are so few rows, it is unlikely anyone
will want to give one up for the status line. Previously, we had
presumed the status line was ON, since that is the default whenever
TELCOM is run.