Add unicode example of all ACS characters
Note that some of them I couldn't pin down an exact shape.

Board of squares I think is a square divided into four pieces.
I approximated it with the Unicode squared location symbol, but it is
not correct.

Lantern symbol nobody alive knows. Some people think it actually is
the Section symbol. However, given that it was at one point
represented by the letter 'I', I'm guessing it is a mathematical phi.

Solid square block is called "square", but I have a hunch it is a full
cursor block.